% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handler.R
\name{handle_xy}
\alias{handle_xy}
\title{Title}
\usage{
handle_xy(input, handler)
}
\arguments{
\item{handler}{}
}
\value{

}
\description{
Title
}
\examples{
df <- data.frame(x = 1:3, y = 2:4)
handle_xy(df, wk::wkt_writer())
}
