% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkmike-package.R
\docType{package}
\name{wkmike-package}
\alias{wkmike}
\alias{wkmike-package}
\title{wkmike: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
